complexVec h and cpp
