+++ include/Titan/vec.h
@@ -27,7 +27,6 @@
 #include <math.h>
 #include <iostream>
 #include <cmath>
-#include <cuComplex.h>
 
 class Vec {
 public:
@@ -66,45 +65,6 @@ public:
     CUDA_CALLABLE_MEMBER double norm() const;
     CUDA_CALLABLE_MEMBER double sum() const;
     CUDA_CALLABLE_MEMBER Vec normalized() const;
-
-private:
-};
-
-class ComplexVec {
-public:
-
-    cuDoubleComplex data[3] = { 0 }; // initialize data to 0
-
-    CUDA_CALLABLE_MEMBER ComplexVec();
-    CUDA_CALLABLE_MEMBER ComplexVec(const ComplexVec & v);
-    CUDA_CALLABLE_MEMBER ComplexVec(cuDoubleComplex x, cuDoubleComplex y, cuDoubleComplex z);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator=(const ComplexVec & v);
-
-    CUDA_CALLABLE_MEMBER ComplexVec & operator+=(const ComplexVec & v);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator-=(const ComplexVec & v);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator*=(const ComplexVec & v);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator/=(const ComplexVec & v);
-
-    CUDA_CALLABLE_MEMBER ComplexVec & operator+=(double x);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator-=(double x);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator*=(double x);
-    CUDA_CALLABLE_MEMBER ComplexVec & operator/=(double x);
-
-    //CUDA_CALLABLE_MEMBER ComplexVec operator-() const; // returns the negative -z = -a -bi
-    CUDA_CALLABLE_MEMBER cuDoubleComplex & operator [] (int n);
-    CUDA_CALLABLE_MEMBER const cuDoubleComplex & operator [] (int n) const;
-    CUDA_CALLABLE_MEMBER friend bool operator==(const ComplexVec & v1, const ComplexVec & v2);
-
-    CUDA_CALLABLE_MEMBER friend ComplexVec operator+(const ComplexVec & v1, const ComplexVec & v2);
-    CUDA_CALLABLE_MEMBER friend ComplexVec operator-(const ComplexVec & v1, const ComplexVec & v2);
-    CUDA_CALLABLE_MEMBER friend ComplexVec operator*(const ComplexVec & v1, const ComplexVec & v2);
-    CUDA_CALLABLE_MEMBER friend ComplexVec operator/(const ComplexVec & v1, const ComplexVec & v2);
-
-    friend std::ostream & operator << (std::ostream & strm, const ComplexVec & v);
-
-    CUDA_CALLABLE_MEMBER void print();
-    //CUDA_CALLABLE_MEMBER double norm() const;
-    CUDA_CALLABLE_MEMBER cuDoubleComplex sum() const;
 
 private:
 };
